% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metrics.R
\name{calc_events}
\alias{calc_events}
\title{Identify and count number of hypoglycaemia or hyperglycaemia events}
\usage{
calc_events(
  data = NULL,
  by_vars,
  threshold = "<70",
  duration = 120,
  event_name = "events"
)
}
\arguments{
\item{data}{A data.table prepared for analysis using `prep_data` command.}

\item{by_vars}{Grouping variables to summarise events by. If missing, assumes all character or factor columns.}

\item{threshold}{Character variable indicating threshold and direction for assigning an event (i.e. <70 or >250).}

\item{duration}{Continuous duration in minutes for time spent above or below threshold to be considered an event.}

\item{event_name}{Character specifying name to be given to new column containing number of events.}
}
\value{
A data table.
}
\description{
Identify and count instances (events) when an individual experience extended hypoglycaemia or hyperglycaemia. I.e. occasions when glucose measurements
  crossed a specified threshold for an extended period of time. The event is only considered to have ended when glucose values return to normal limits for 15 minutes or more.
  Missing data is ignored if lasting 10 minutes or less, otherwise it interrupts the event.
  Very long extended events may be counted as multiple consecutive events.
}
