% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metrics.R
\name{calc_percent}
\alias{calc_percent}
\title{Convert counts to percentages}
\usage{
calc_percent(data, cols = NULL, denom = NULL)
}
\arguments{
\item{data}{Data table with columns to convert to percentages}

\item{cols}{Character vector of names of columns containing counts to convert to percentages. If not supplied, columns containing the text 'tir', 'tbr', 'tar', or 'ttr' are converted.}

\item{denom}{Character name of column containing denominator counts, or integer vector of counts to use as a denominator. If not supplied, 'obs_n' is used.}
}
\value{
Data table is modified in place, the function does not copy or return the data table.
}
\description{
Convert counts to percentages
}
