% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metrics.R
\name{calc_summaries}
\alias{calc_summaries}
\title{Calculate continuous CGM summary metrics}
\usage{
calc_summaries(data, by_vars = NULL)
}
\arguments{
\item{data}{A data.table prepared for analysis using `prep_data` command.}

\item{by_vars}{Grouping variables to summarise statistics by. If missing, assumes all character or factor columns.}
}
\value{
A data.table containing the CGM summary metrics in long form.
}
\description{
Calculates summary CGM metrics including:
* Proportions of times spent in, above, or below, consensus ranges.
* Summary measures (mean, standard deviation, coefficient of variation).
}
