% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metrics.R
\name{calc_times}
\alias{calc_times}
\title{Convert percentages to hours and minutes}
\usage{
calc_times(data, cols = NULL, hours = 24)
}
\arguments{
\item{data}{Data table with columns to convert from percentages to hours and minutes.}

\item{cols}{Character vector of names of columns containing percentages to convert to times. If not supplied, columns containing the text 'tir', 'tbr', 'tar', or 'ttr' are converted.}

\item{hours}{Integer denoting maximum number of hours. Default is 24 hours.}
}
\value{
Data table is modified in place, the function does not copy or return the data table.
}
\description{
Convert percentages to hours and minutes
}
