% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGMmetrics-package.R
\docType{package}
\name{CGMmetrics-package}
\alias{CGMmetrics}
\alias{CGMmetrics-package}
\title{CGMmetrics: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vsritter/CGMmetrics}
  \item Report bugs at \url{https://github.com/vsritter/CGMmetrics/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
