% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{completeDT}
\alias{completeDT}
\title{Complete a data table with missing combinations of data}
\usage{
completeDT(DT, cols, fill = NULL)
}
\arguments{
\item{DT}{A data.table.}

\item{cols}{Vector of column names to be completed. Date variables are filled in as a full sequence from first to last.}

\item{fill}{A named list that for each variable supplies a single value to use instead of NA for missing combinations.}
}
\value{
A `data.table`
}
\description{
Turns implicit missing values into explicit missing values by adding empty rows. A simpler data.table version of `tidyr::complete()`. Modified from post on stackoverflow (https://stackoverflow.com/a/43483804/4241780).
}
