% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metrics.R
\name{make_metrics}
\alias{make_metrics}
\title{Calculate core and secondary CGM endpoints.}
\usage{
make_metrics(data, id_vars, time_var = NULL, output = NULL)
}
\arguments{
\item{data}{A data.table prepared for analysis using `prep_data` command.}

\item{id_vars}{Additional subject level variables to include in the final dataset.}

\item{time_var}{Time variables to group and summarise statistics by. If missing, assumes all character or factor columns.}

\item{output}{What type of dataset of output. 'Wide' (default) or 'long'.}
}
\value{
A data.table with metrics named with the prefix 'cgm'.
}
\description{
Summarise continuous glucose monitoring (CGM) data to produce core and secondary continuous endpoints
  as specified in *Continuous glucose monitoring and metrics for clinical trials: an intertional consensus statement* by
  Bettelino *et al* (2023).
}
