% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare raw CGM data for analysis}
\usage{
prep_data(
  data,
  id_vars = NULL,
  var_datetime = NULL,
  var_glucose = NULL,
  var_keep = NULL,
  tz = Sys.timezone(),
  accuracy = 300L
)
}
\arguments{
\item{data}{A data.frame or data.table with at least two columns indicating the datetime and glucose reading.}

\item{id_vars}{A character vector of the names of any grouping variables to keep in the dataset.}

\item{var_datetime}{The name of the datetime column. If missing, assumes the second last column.}

\item{var_glucose}{The name of the column containing the glucose measurements. If missing, assumes the last column.}

\item{var_keep}{The names of any other columns to keep in the final dataset.}

\item{tz}{The timezone of the datetime field. Defaults to the system timezone if not specified.}

\item{accuracy}{Interval in seconds to round time to, defaults to 5 minutes (300 seconds).}
}
\value{
A data.table with the following columns; grouping variables (as indicated), datetime variables (obs_dttm, obt_dttmr, obs_dt, obs_tm), glucose measurements (glu).
}
\description{
Edit a simple dataset contain grouping variables, a datetime variable, and glucose measurements by:
* Rounding times down to the closest 5 minute interval.
* Splitting the datetime variable into separate date and time variables.
* Adding extra rows to indicate missing observations, so each day has 288 observations and all days are included between the first and last observed day.
* Adding an indicator of daytime vs nocturnal observations.
}
