% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metrics.R
\name{make_metrics}
\alias{make_metrics}
\title{Calculate core and secondary CGM endpoints.}
\usage{
make_metrics(data, by_vars, longform = FALSE)
}
\arguments{
\item{data}{A data.table prepared for analysis using `prep_data` command.}

\item{by_vars}{Grouping variables to summarise statistics by. If missing, assumes all character or factor columns.}

\item{longform}{Logical indicating whether to reshape data long, or leave in wide form (the default).}
}
\value{
A data.table with metrics named with the prefix 'cgm'.
}
\description{
Summarise continuous glucose monitoring (CGM) data to produce core and secondary continuous endpoints
  as specified in *Continuous glucose monitoring and metrics for clinical trials: an intertional consensus statement* by
  Bettelino *et al* (2023).
}
